# api/openapi.yaml
openapi: 3.0.0
info:
  title: Arth-Arbitrage
  version: 0.0.1
  description: Arth Arbitrage API
  termsOfService: MIT
servers:
  - url: 'http://localhost:4000'
    variables: {}
    description: arth-arbitrage end-point
paths:
  /exchanges:
    summary: exchange list
    description: List of supported exchanges
    get:
      summary: Get exchange list
      description: Get list of exchanges
      operationId: listExchanges
      responses:
        '200':
          description: A paged array of exchanges
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exchanges'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/exchanges/{id}':
    get:
      summary: Exchange details
      description: Details of the exchange
      operationId: getExchange
      parameters:
        - name: id
          in: path
          description: ID of anaytics exchange to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: exchange response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exchange'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/exchanges/{id}/pools':
    get:
      summary: Pool List
      description: List of liquidity pools
      operationId: getSwapPools
      parameters:
        - name: id
          in: path
          description: ID of exchange
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: exchange response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwapPools'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
  /lenders:
    summary: Lending service
    description: List of Lending service
    get:
      summary: Lending service list
      description: List of lending services
      operationId: listLenders
      responses:
        '200':
          description: A paged array of lending services
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lenders'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/lenders/{id}':
    get:
      summary: Get lender
      description: Returns a details of a lender
      operationId: getLender
      parameters:
        - name: id
          in: path
          description: ID of lender
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: lender response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lender'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/lenders/{id}/pools':
    get:
      summary: Pool List
      description: List of liquidity pools
      operationId: getLenderPools
      parameters:
        - name: id
          in: path
          description: ID of exchange
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: exchange response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LenderPools'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/swap':
    post:
      summary: process swap 
      description: Process swap transaction
      operationId: swap
      requestBody:
        description: Swap to be perfomed
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Swap'
      responses:
        '201':
          description: Created              
        '422':
          description: Failed     
  '/multiswap':
    post:
      summary: process swap 
      description: Process swap transaction
      operationId: multiswap
      requestBody:
        description: Swap to be perfomed
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiSwap'
      responses:
        '201':
          description: Created              
        '422':
          description: Failed    
          
components:
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string

    Exchanges:
      type: array
      items:
        $ref: '#/components/schemas/Exchange'
    Exchange:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SwapPools:
      type: array
      items:
        $ref: '#/components/schemas/SwapPool'
    SwapPool:
      type: object
      required:
        - id
        - name
        - token0
        - token1
        - reserve0
        - reserve1
        - price0
        - price1
        - fees
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        token0:
          type: string
        token1:
          type: string
        reserve0:
          type: number
          format: float
        reserve1:
          type: number
          format: float
        price0:
          type: number
          format: float
        price1:
          type: number
          format: float
        fees:
          type: number
          format: float
          
    Lenders:
      type: array
      items:
        $ref: '#/components/schemas/Lender'
    Lender:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    LenderPools:
      type: array
      items:
        $ref: '#/components/schemas/LenderPool'
    LenderPool:
      type: object
      required:
        - id
        - name
        - token
        - reserve
        - fees
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        token:
          type: string
        reserve:
          type: number
          format: float
        fees:
          type: number
          format: float
    Swap:
      type: object
      required:
        - amount
        - token0
        - token1
        - exchange
      properties:
        token0:
          type: string
        token1:
          type: string
        amount:
          type: number
          format: float
        exchange:
          type: integer
          format: int64
    MultiSwap:
      type: object
      properties:
        lender:
          type: integer
          format: int64
        swaps:
          type: array
          items:
            $ref: '#/components/schemas/Swap'